Hidden functions (analysis)

Functions for the palasso manuscript.

.prepare(X, cutoff = NULL, quantile = NULL, scale = TRUE)

.simulate(x, effects)

.predict(y, X, nfolds.ext = 5, nfolds.int = 5, ...)

.select(y, X, index, family = "binomial", nfolds = 5, ...)

Arguments

X
covariates: matrix with \(n\) rows and \(p\) columns

x
covariates: list of length \(k\), including matrices with \(n\) rows and \(p\) columns

effects
number of causal covariates: vector of length \(k\)

y
response: vector of length \(n\)

nfolds.ext
number of external folds

...
arguments for palasso

index
indices of causal covariates: list of length \(k\), including vectors

trial
development option

Value

to do

Details

The function .prepare pre-processes sequencing data. It removes features with a low total abundance, adjusts for different library sizes, binarises the counts with cutoff zero or takes the Anscombe transform, and scales all covariates to mean zero and unit variance. (If the cutoff differed from zero, we would have to first normalise, then binarise, and finally filter the raw counts.)

The function .simulate simulates a response. Exploiting an experimental covariate matrix, it allows for different numbers of non-zero coefficients for X and Z.

The function .predict estimates the predictive performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z). Minimising the deviance, it returns the deviance and the area under the curve. Currently, only the binomial family is implemented.

The function .select estimates the selective performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z). Limiting the number of covariates to \(10\), it returns the number of selected covariates, and the number of correctly selected covariates.

See also

Use palasso to fit the paired lasso.

Examples

set.seed(1)
n <- 50; p <- 100
X <- matrix(rpois(n*p,lambda=4),nrow=n,ncol=p)
x <- palasso:::.prepare(X)
y <- palasso:::.simulate(x,effects=c(1,2))
palasso:::.predict(y,x)

#> $info
#>   nfolds.ext nfolds.int          time
#> 1          5          5 4.209421 secs
#> 
#> $deviance
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5     1.390625   1.395063    1.390625   1.325492   1.403424    1.325492
#> 10    1.450516   1.400899    1.470338   1.500959   1.407533    1.495150
#> 20    1.467288   1.397189    1.488180   1.527195   1.416801    1.517447
#> Inf   1.467288   1.398821    1.488180   1.527195   1.517579    1.517447
#>     between_xz within_xz   paired   trial1   trial2   trial3   trial4
#> 5     1.390625  1.398262 1.367785 1.397774 1.367785 1.406872 1.363852
#> 10    1.450516  1.424210 1.596432 1.472129 1.596432 1.474732 1.423742
#> 20    1.467288  1.441167 1.645832 1.487795 1.645832 1.487795 1.440514
#> Inf   1.467288  1.441167 1.718443 1.489427 1.718443 1.489427 1.440514
#> 
#> $auc
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5    0.5264423  0.4423077   0.5264423  0.6682692  0.4711538   0.6682692
#> 10   0.5520833  0.4775641   0.5456731  0.6602564  0.4855769   0.6554487
#> 20   0.5520833  0.4839744   0.5440705  0.6778846  0.5192308   0.6698718
#> Inf  0.5520833  0.4887821   0.5440705  0.6778846  0.5568910   0.6698718
#>     between_xz within_xz    paired    trial1    trial2    trial3    trial4
#> 5    0.5264423 0.6290064 0.6089744 0.5440705 0.6089744 0.5456731 0.6282051
#> 10   0.5520833 0.6049679 0.6250000 0.5633013 0.6250000 0.5008013 0.6538462
#> 20   0.5520833 0.6033654 0.6153846 0.5008013 0.6153846 0.5008013 0.6538462
#> Inf  0.5520833 0.6033654 0.5913462 0.5056090 0.5913462 0.5056090 0.6538462
#> 
#> $mse
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5    0.2511776  0.2521719   0.2511776  0.2342846  0.2539897   0.2342846
#> 10   0.2641671  0.2535401   0.2681709  0.2529158  0.2540529   0.2525655
#> 20   0.2675057  0.2526740   0.2712718  0.2554788  0.2569348   0.2549654
#> Inf  0.2675057  0.2527902   0.2712718  0.2554788  0.2598926   0.2549654
#>     between_xz within_xz    paired    trial1    trial2    trial3    trial4
#> 5    0.2511776 0.2526642 0.2446934 0.2529179 0.2446934 0.2551898 0.2445111
#> 10   0.2641671 0.2576737 0.2757328 0.2693563 0.2757328 0.2700286 0.2575006
#> 20   0.2675057 0.2605593 0.2857310 0.2725010 0.2857310 0.2725010 0.2608392
#> Inf  0.2675057 0.2605593 0.2906830 0.2726172 0.2906830 0.2726172 0.2608392
#> 
#> $mae
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5    0.4957308  0.5012405   0.4957308  0.4624717  0.5011554   0.4624717
#> 10   0.5011887  0.5017250   0.5048246  0.4286383  0.4956691   0.4331166
#> 20   0.5030825  0.5002803   0.5064563  0.4225268  0.4915612   0.4272417
#> Inf  0.5030825  0.4994964   0.5064563  0.4225268  0.4645089   0.4272417
#>     between_xz within_xz    paired    trial1    trial2    trial3    trial4
#> 5    0.4957308 0.5002270 0.4740567 0.4990318 0.4740567 0.5014303 0.4885961
#> 10   0.5011887 0.5003498 0.4655633 0.5093666 0.4655633 0.5099837 0.4940540
#> 20   0.5030825 0.5009838 0.4742355 0.5104327 0.4742355 0.5104327 0.4959478
#> Inf  0.5030825 0.5009838 0.4648347 0.5096488 0.4648347 0.5096488 0.4959478
#> 
#> $class
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5         0.44       0.52        0.44       0.38       0.54        0.38
#> 10        0.40       0.50        0.40       0.28       0.52        0.28
#> 20        0.40       0.50        0.40       0.28       0.50        0.28
#> Inf       0.40       0.50        0.40       0.28       0.50        0.28
#>     between_xz within_xz paired trial1 trial2 trial3 trial4
#> 5         0.44      0.42   0.44   0.48   0.44   0.52   0.38
#> 10        0.40      0.40   0.34   0.44   0.34   0.48   0.34
#> 20        0.40      0.40   0.38   0.48   0.38   0.48   0.34
#> Inf       0.40      0.40   0.44   0.48   0.44   0.48   0.34
#> 

palasso:::.select(y,x,attributes(y))

#> $shots
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5            4          2           4          5          2           5
#> 10           6          2           4          9          2           9
#> 20           6          2           4          9         12           9
#> Inf          6          2           4          9         12           9
#>     between_xz within_xz paired trial1 trial2 trial3 trial4
#> 5            0         0      5      2      5      2      0
#> 10           6         0      9      2      9      2      6
#> 20           6         0      9      2      9      2      6
#> Inf          6         0      9      2      9      2      6
#> 
#> $hits
#>     standard_x standard_z standard_xz adaptive_x adaptive_z adaptive_xz
#> 5            0          0           0          0          0           0
#> 10           0          0           0          1          0           1
#> 20           0          0           0          1          0           1
#> Inf          0          0           0          1          0           1
#>     between_xz within_xz paired trial1 trial2 trial3 trial4
#> 5            0         0      0      0      0      0      0
#> 10           0         0      1      0      1      0      0
#> 20           0         0      1      0      1      0      0
#> Inf          0         0      1      0      1      0      0
#>