Methods for class "palasso"

This page lists the main methods for class "palasso".

# S3 method for class 'palasso'
predict(object, newdata, model = "paired", s = "lambda.min", max = NULL, ...)

# S3 method for class 'palasso'
coef(object, model = "paired", s = "lambda.min", max = NULL, ...)

# S3 method for class 'palasso'
weights(object, model = "paired", max = NULL, ...)

# S3 method for class 'palasso'
fitted(object, model = "paired", s = "lambda.min", max = NULL, ...)

# S3 method for class 'palasso'
residuals(object, model = "paired", s = "lambda.min", max = NULL, ...)

# S3 method for class 'palasso'
deviance(object, model = "paired", max = NULL, ...)

# S3 method for class 'palasso'
logLik(object, model = "paired", max = NULL, ...)

# S3 method for class 'palasso'
summary(object, model = "paired", ...)

Arguments

object

palasso object

newdata

covariates: list of matrices, each with \(n\) rows (samples) and \(p\) columns (variables)

model

character "paired", or an entry of names(object)

s

penalty parameter: character "lambda.min" or "lambda.1se", positive numeric, or NULL (entire sequence)

max

maximum number of non-zero coefficients, positive integer, or NULL

...

further arguments for predict.cv.glmnet, coef.cv.glmnet, or deviance.glmnet

Details

By default, the function predict returns the linear predictor (type="link"). Consider predicting the response (type="response").

See also