Analysis functions for manuscript

Functions for the palasso manuscript.

.prepare(X, filter = 1, cutoff = "zero", scale = TRUE)

.simulate(x, effects)

.predict(
  y,
  X,
  nfolds.ext = 5,
  nfolds.int = 5,
  adaptive = TRUE,
  standard = TRUE,
  elastic = TRUE,
  shrink = TRUE,
  family = "binomial",
  ...
)

.select(y, X, index, nfolds = 5, standard = TRUE, adaptive = TRUE, ...)

Arguments

X
covariates: matrix with \(n\) rows and \(p\) columns

filter
numeric, multiplying the sample size

cutoff
character "zero", "knee", or "half"

scale
logical

x
covariates: list of length \(k\), including matrices with \(n\) rows and \(p\) columns

effects
number of causal covariates: vector of length \(k\)

y
response: vector of length \(n\)

nfolds.ext
number of external folds

...
arguments for palasso

index
indices of causal covariates: list of length \(k\), including vectors

trial
development option

Details

.prepare: pre-processes sequencing data by removing features with a low total abundance, and adjusting for different library sizes; obtains two transformations of the same data by (1) binarising the counts with some cutoff and (2) taking the Anscombe transform; scales all covariates to mean zero and unit variance.

.simulate: simulates the response by exploiting two experimental covariate matrices; allows for different numbers of non-zero coefficients for X and Z.

.predict: estimates the predictive performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z); minimises the loss function "deviance", but also returns other loss functions; supports logistic and Cox regression.

.select: estimates the selective performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z); limits the number of covariates to \(10\); returns the number of selected covariates, and the number of correctly selected covariates.

See also

Examples

if (FALSE) set.seed(1)
n <- 30; p <- 40
X <- matrix(rpois(n*p,lambda=3),nrow=n,ncol=p)
x <- palasso:::.prepare(X)
y <- palasso:::.simulate(x,effects=c(1,2))
predict <- palasso:::.predict(y,x)

#> Warning: Install "CorShrink" from CRAN archive, or shrink=FALSE.

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#>                        lambda      nzero length
#> standard_x  [0.3,(0.1),0.003] [0,(5),12]    100
#> standard_z  [0.2,(0.2),0.002] [0,(1),18]    100
#> standard_xz [0.3,(0.1),0.003] [0,(6),18]    100
#> adaptive_x  [0.7,(0.7),0.007]  [0,(0),9]    100
#> adaptive_z  [0.2,(0.1),0.002] [0,(1),19]    100
#> adaptive_xz [0.9,(0.9),0.009] [0,(0),12]    100
#> between_xz  [0.4,(0.4),0.004] [0,(0),14]    100
#> within_xz        [3,(3),0.03] [0,(0),11]    100
#> elastic95   [0.3,(0.1),0.003] [0,(6),19]    200

#> Warning: Install "CorShrink" from CRAN archive, or shrink=FALSE.

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#>                         lambda       nzero length
#> standard_x  [0.3,(0.08),0.003]  [0,(8),14]    100
#> standard_z   [0.2,(0.2),0.002]  [0,(0),20]    100
#> standard_xz  [0.3,(0.1),0.003]  [0,(8),14]    100
#> adaptive_x      [4,(0.6),0.04]  [0,(6),11]    100
#> adaptive_z   [0.2,(0.2),0.002]  [0,(0),16]    100
#> adaptive_xz     [5,(0.5),0.05]  [0,(8),10]    100
#> between_xz      [1,(0.1),0.01] [0,(11),11]    100
#> within_xz        [500,(500),5]  [0,(1),11]    100
#> elastic95   [0.4,(0.07),0.004] [0,(10),16]    200

#> Warning: Install "CorShrink" from CRAN archive, or shrink=FALSE.

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#>                        lambda      nzero length
#> standard_x  [0.2,(0.2),0.002] [0,(0),18]    100
#> standard_z  [0.1,(0.1),0.001] [0,(0),23]    100
#> standard_xz [0.2,(0.2),0.002] [0,(0),20]    100
#> adaptive_x       [1,(1),0.01] [0,(0),11]    100
#> adaptive_z  [0.1,(0.1),0.001] [0,(0),18]    100
#> adaptive_xz      [1,(1),0.01] [0,(0),15]    100
#> between_xz  [0.6,(0.3),0.006] [0,(7),13]    100
#> within_xz       [20,(20),0.2] [0,(0),11]    100
#> elastic95   [0.3,(0.3),0.003] [0,(0),23]    200

#> Warning: Install "CorShrink" from CRAN archive, or shrink=FALSE.

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#>                         lambda      nzero length
#> standard_x  [0.3,(0.08),0.003] [0,(7),13]    100
#> standard_z   [0.2,(0.2),0.002] [0,(0),19]    100
#> standard_xz  [0.3,(0.1),0.003] [0,(6),14]    100
#> adaptive_x      [4,(0.1),0.04]  [0,(6),8]    100
#> adaptive_z   [0.2,(0.2),0.002] [0,(0),17]    100
#> adaptive_xz       [4,(2),0.04]  [0,(4),8]    100
#> between_xz   [0.8,(0.1),0.008] [0,(6),10]    100
#> within_xz       [30,(0.3),0.3]  [0,(7),8]    100
#> elastic95    [0.3,(0.1),0.003] [0,(5),17]    200

#> Warning: Install "CorShrink" from CRAN archive, or shrink=FALSE.

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#> Warning: one multinomial or binomial class has fewer than 8  observations; dangerous ground

#>                        lambda       nzero length
#> standard_x  [0.3,(0.1),0.003]  [0,(8),15]    100
#> standard_z  [0.2,(0.2),0.002]  [0,(2),24]    100
#> standard_xz [0.3,(0.1),0.003] [0,(10),18]    100
#> adaptive_x     [1,(0.7),0.01]   [0,(1),8]    100
#> adaptive_z  [0.2,(0.2),0.002]  [0,(2),18]    100
#> adaptive_xz    [2,(0.8),0.02]  [0,(1),11]    100
#> between_xz  [0.6,(0.3),0.006]  [0,(2),12]    100
#> within_xz        [7,(4),0.07]   [0,(1),9]    100
#> elastic95   [0.3,(0.1),0.003] [0,(11),23]    200

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

select <- palasso:::.select(y,x,attributes(y))