Other functions (manuscript)

Functions for the palasso manuscript.

.prepare(X, filter = 1, cutoff = "zero", scale = TRUE)

.simulate(x, effects)

.predict(y, X, nfolds.ext = 5, nfolds.int = 5, standard = TRUE,
  family = "binomial", ...)

.select(y, X, index, nfolds = 5, standard = TRUE, adaptive = TRUE, ...)

Arguments

X
covariates: matrix with \(n\) rows and \(p\) columns

filter
numeric, multiplying the sample size

cutoff
character "zero", "knee", or "half"

scale
logical

x
covariates: list of length \(k\), including matrices with \(n\) rows and \(p\) columns

effects
number of causal covariates: vector of length \(k\)

y
response: vector of length \(n\)

nfolds.ext
number of external folds

...
arguments for palasso

index
indices of causal covariates: list of length \(k\), including vectors

trial
development option

Value

to do

Details

The function .prepare pre-processes sequencing data. It removes features with a low total abundance, adjusts for different library sizes, binarises the counts with cutoff zero or takes the Anscombe transform, and scales all covariates to mean zero and unit variance. (If the cutoff differed from zero, we would have to first normalise, then binarise, and finally filter the raw counts.)

The function .simulate simulates a response. Exploiting an experimental covariate matrix, it allows for different numbers of non-zero coefficients for X and Z.

The function .predict estimates the predictive performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z). Minimising the deviance, it returns the deviance and the area under the curve. Currently, only the binomial family is implemented.

The function .select estimates the selective performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z). Limiting the number of covariates to \(10\), it returns the number of selected covariates, and the number of correctly selected covariates.

See also

Examples

set.seed(1)
n <- 30; p <- 40
X <- matrix(rpois(n*p,lambda=3),nrow=n,ncol=p)
x <- palasso:::.prepare(X)
y <- palasso:::.simulate(x,effects=c(1,2))
predict <- palasso:::.predict(y,x)
select <- palasso:::.select(y,x,attributes(y))