Analysis functions for manuscript

Functions for the palasso manuscript.

.prepare(X, filter = 1, cutoff = "zero", scale = TRUE)

.simulate(x, effects)

.predict(y, X, nfolds.ext = 5, nfolds.int = 5, adaptive = TRUE,
  standard = TRUE, family = "binomial", ...)

.select(y, X, index, nfolds = 5, standard = TRUE, adaptive = TRUE,
  ...)

Arguments

X
covariates: matrix with \(n\) rows and \(p\) columns

filter
numeric, multiplying the sample size

cutoff
character "zero", "knee", or "half"

scale
logical

x
covariates: list of length \(k\), including matrices with \(n\) rows and \(p\) columns

effects
number of causal covariates: vector of length \(k\)

y
response: vector of length \(n\)

nfolds.ext
number of external folds

...
arguments for palasso

index
indices of causal covariates: list of length \(k\), including vectors

trial
development option

Details

.prepare: pre-processes sequencing data by removing features with a low total abundance, and adjusting for different library sizes; obtains two transformations of the same data by (1) binarising the counts with some cutoff and (2) taking the Anscombe transform; scales all covariates to mean zero and unit variance.

.simulate: simulates the response by exploiting two experimental covariate matrices; allows for different numbers of non-zero coefficients for X and Z.

.predict: estimates the predictive performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z); minimises the loss function "deviance", but also returns other loss functions; supports logistic and Cox regression.

.select: estimates the selective performance of different lasso models (standard X and/or Z, adaptive X and/or Z, paired X and Z); limits the number of covariates to \(10\); returns the number of selected covariates, and the number of correctly selected covariates.

See also

Examples

set.seed(1)
n <- 30; p <- 40
X <- matrix(rpois(n*p,lambda=3),nrow=n,ncol=p)
x <- palasso:::.prepare(X)
y <- palasso:::.simulate(x,effects=c(1,2))
predict <- palasso:::.predict(y,x)

#> Registered S3 method overwritten by 'R.oo':
#>   method        from       
#>   throw.default R.methodsS3

#>                         lambda       nzero length
#> standard_x   [0.3,(0.3),0.003]  [0,(0),14]    100
#> standard_z  [0.2,(0.09),0.002] [0,(10),20]    100
#> standard_xz [0.3,(0.05),0.003] [0,(11),14]    100
#> adaptive_x   [0.1,(0.1),0.001]  [0,(1),14]    100
#> adaptive_z  [0.1,(0.05),0.001] [0,(10),20]    100
#> adaptive_xz [0.3,(0.04),0.003] [0,(12),14]    100
#> between_xz   [0.3,(0.3),0.003]  [0,(0),14]    100
#> within_xz   [0.3,(0.04),0.003] [0,(12),14]    100
#>                         lambda       nzero length
#> standard_x   [0.2,(0.2),0.002]  [0,(1),18]    100
#> standard_z  [0.3,(0.04),0.003] [0,(12),17]    100
#> standard_xz  [0.3,(0.2),0.003]  [0,(3),23]    100
#> adaptive_x   [0.1,(0.1),0.001]  [0,(1),18]    100
#> adaptive_z  [0.1,(0.02),0.001] [0,(11),17]    100
#> adaptive_xz  [0.3,(0.2),0.003]  [0,(3),23]    100
#> between_xz   [0.3,(0.2),0.003]  [0,(3),23]    100
#> within_xz    [0.3,(0.2),0.003]  [0,(3),23]    100
#>                         lambda       nzero length
#> standard_x   [0.2,(0.2),0.002]  [0,(0),14]    100
#> standard_z  [0.3,(0.03),0.003] [0,(14),15]    100
#> standard_xz [0.3,(0.09),0.003] [0,(10),17]    100
#> adaptive_x   [0.1,(0.1),0.001]  [0,(0),14]    100
#> adaptive_z  [0.1,(0.01),0.001] [0,(14),15]    100
#> adaptive_xz [0.3,(0.09),0.003] [0,(10),17]    100
#> between_xz  [0.3,(0.09),0.003] [0,(10),17]    100
#> within_xz   [0.3,(0.09),0.003] [0,(10),17]    100
#>                         lambda       nzero length
#> standard_x   [0.3,(0.3),0.003]  [0,(0),17]    100
#> standard_z  [0.2,(0.04),0.002] [0,(15),17]    100
#> standard_xz  [0.3,(0.3),0.003]  [0,(0),20]    100
#> adaptive_x  [0.1,(0.07),0.001]  [0,(4),17]    100
#> adaptive_z  [0.1,(0.02),0.001] [0,(15),17]    100
#> adaptive_xz  [0.3,(0.1),0.003]  [0,(5),20]    100
#> between_xz   [0.3,(0.3),0.003]  [0,(0),20]    100
#> within_xz    [0.3,(0.1),0.003]  [0,(5),20]    100
#>                         lambda       nzero length
#> standard_x  [0.3,(0.06),0.003] [0,(13),17]    100
#> standard_z  [0.2,(0.03),0.002] [0,(14),17]    100
#> standard_xz [0.3,(0.07),0.003] [0,(13),16]    100
#> adaptive_x  [0.1,(0.03),0.001] [0,(13),17]    100
#> adaptive_z  [0.1,(0.01),0.001] [0,(14),17]    100
#> adaptive_xz [0.3,(0.07),0.003] [0,(13),16]    100
#> between_xz  [0.3,(0.07),0.003] [0,(13),16]    100
#> within_xz   [0.3,(0.07),0.003] [0,(13),16]    100

#> Registered S3 methods overwritten by 'ggplot2':
#>   method         from 
#>   [.quosures     rlang
#>   c.quosures     rlang
#>   print.quosures rlang

select <- palasso:::.select(y,x,attributes(y))