Paired lasso

The function palasso fits the paired lasso. Use this regression technique if the covariates are numerous and occur in pairs.

palasso(y, X, trial = FALSE, ...)

Arguments

y
response: vector of length \(n\),

X
covariates: list of matrices, each with \(n\) rows (samples) and \(p\) columns (variables)

trial
development option

...
further arguments for cv.glmnet or glmnet

Value

This function returns an object of class palasso. Available methods include predict, coef, weights, fitted, residuals, deviance, logLik, and summary. Click here for hidden functions.

Details

Let x denote one entry of the list X. See glmnet for alternative specifications of y and x. Among the further arguments, family must equal "gaussian", "binomial", "poisson", or "cox", and penalty.factor must not be used. This method (hopefully) improves prediction if alpha\(=1\) (lasso) and pmax\(<<p\) (non-zero coefficients).

References

A Rauschenberger, RX Menezes, MA Jonker, and MA van de Wiel (2018). "Penalised regression with paired covariates." Manuscript in preparation. a.rauschenberger@vumc.nl

Examples

set.seed(1)
n <- 40; p <- 1000
y <- rbinom(n=n,size=1,prob=0.5)
X <- lapply(1:2,function(x) matrix(rnorm(n*p),nrow=n,ncol=p))
object <- palasso(y=y,X=X,family="binomial",pmax=10)